update all id to uuid()

+++ database/migrations/1750759289851_create_update_users_add_extra_fields_table.ts
@@ -4,18 +4,22 @@ export default class UpdateUsersSpecialtyField extends BaseSchema {
   protected tableName = 'users'
 
   public async up() {
-    this.schema.alterTable(this.tableName, (table) => {
-      // Supprimer colonne mal nommée si elle existe
-      table.dropColumn('speciality') // pas de `.catch` ici, Adonis ne supporte pas
-      // Ajouter la colonne 'specialty' si elle n'existe pas
-      table.string('specialty').nullable()
-    })
+    const hasSpecialty = await this.db.schema.hasColumn(this.tableName, 'specialty')
+
+    if (!hasSpecialty) {
+      await this.schema.alterTable(this.tableName, (table) => {
+        table.string('specialty').nullable()
+      })
+    }
   }
 
   public async down() {
-    this.schema.alterTable(this.tableName, (table) => {
-      table.dropColumn('specialty')
-      table.string('speciality').nullable()
-    })
+    const hasSpecialty = await this.db.schema.hasColumn(this.tableName, 'specialty')
+
+    if (hasSpecialty) {
+      await this.schema.alterTable(this.tableName, (table) => {
+        table.dropColumn('specialty')
+      })
+    }
   }
 }
